Added nix support
The files are taken from plutus-starter without change.

They are sufficient to:
1. start a nix-shell in the project directory
2. Support a build that works with vscode and the included devcontainer.

They are not sufficient by themselves to get vscode working 'usefully'
without the devcontainer.

Currently 'plutus-starter' appears in various places.
This should be replaced by 'quadraticvoting' as appropriate. i.e. the
files should be specialized (further) to the needs of this project.