Remove Support For Multiple UTxOs
Without Vasil features, spending a UTxO from a script address requires
the transaction to provide the raw script for that UTxO. Distribution
endpoint with multiple UTxOs requires the presence of all the UTxOs.
Which means the transaction needs to provide the script for every
single one of them. Meaning the transaction size can easily surpass
the limit.

This commit limits the minting script to be able to validate singular
asset mints, and also reduces on-chain validator's size by removing
support for merging multiple datums, etc.

The bash scripts are NOT updated to match these changes yet.