Add Distribution Support to `qvf-cli`.
The `qvf-cli` application can now produce `--tx-out` arguments to be
passed to `cardano-cli transaction build` command for triggering the
prize distribution of the contract.

There is also the added utility of getting the equivalent slot number
of a deadline specified in a provided datum.

The issue with on-chain deadline checks should also be resolved.