Get SAM build and SAM local invoke/start-api working
However this is a temporary solution. It only works for the static
build.

The dynamic build suffers from the 'dependency' problem.

However I've found an article
(https://www.tweag.io/blog/2020-09-30-bazel-static-haskell/) which might
offer a solution which kills all our outstanding 'birds with one stone'.

In case you we're wondering - all of the other 'simple' solutions to
getting the static build don't work.

But the one above looks promising. It looks like there is a musl ghc
installation (pre-built) in a nixpkgs revision. If that pans out (i.e.
there is such an installation and I can get nix to dtrt) then we're
pretty much done :-) We have a single build that works for dev (SAM) and
will work as a github action, and it doesn't involve Docker or external
runners or any of that stuff.