Further Progress.
For now, every project UTxO must first turn into an "escrow" UTxO before
getting permission for having their tokens burnt. This doesn't seem to
be efficient as the projects that don't reach their requested fund will
be subject to one extra transaction fee in order to have their
registration fees refunded.

This inefficiency however, is allowing a much cleaner logic and
codebase. And also, with the intention of bringing the contract to a
testable state as soon as possible, postponing the optimization of this
part of the flow seems to be a reasonable approach.