Split the Authentication Token Name in Two
To simplify the off-chain transaction construction, the authentication
asset for the deadline UTxO, now has a different token name compared to
the UTxO which keeps the record of registered projects (the main asset).

Additionally, the main asset now has an empty token name.

Another change, is the attachment of a minimal value as an inline datum
to the minting script reference UTxOs for a more robust and convenient
distinction between the two.