Restrict escrow UTxOs to key holder.
Since, at this point, we're not utilizing the implemented functionality
of the escrow UTxOs, leaving them open to the respective project owners
(as they should be) allows projects owners to simply mark those UTxOs
for another one of their own wallets and easily extract all the
Lovelaces.

By requiring the key holder's signature (which most of our interactions
in this version already do) we can cover this security flaw.

Some of the scripts have also undergone minor changes (re-ordering of
their inputs).