Turning the Deadline Milliseconds Into an Argument
Previously, the deadline was defined as a variable inside the initiation
script, which resulted in merge conflicts and implicit problems. This PR
aims to resolve this issue by turning the deadline into an argument
for the `initiate_fund` function.

Furthermore, two additional helper functions are defined for
convenience: `posix_one_week_from_now` and `posix_one_month_from_now`,
which return the POSIX time in milliseconds one week (and one month) from
the moment of invocation.