Fix the JSON Injection of a New Registeration
Previously, injection of a newly registered project into the
corresponding JSON file on disk occurred after a very strict check: it
only looked for the presence of another project with the exact object.
This proved problematic as there is still no off-chain error handling.

This commits resolves this issue by making the check exclusive to the
token name of previously registered projects. That is, if an object with
the same token name is found, it is replaced by the new project's
object.