Add a boolean flag to project removal redeemer.
Currently there inconsistency among various off-chain solutions, some of
them produce the change output at index 0, while others produce them as
the last index.

This flag allows off-chain to indicate where it intends to produce the
change output, so that the validator knows where to look for it.
Previously the script expected the change output to sit at index 0.